requirements.txt
Flask~=1.1

################################################################
#
# Python package requirements for the TensorFlow Tutorials:
# https://github.com/Hvass-Labs/TensorFlow-Tutorials
#
# If you are using Anaconda then you can install all required
# Python packages by running the following commands in a shell:
#
#     conda create --name tf python=3
#     activate tf
#     pip install -r requirements.txt
#
# Note that you have to edit this file to select whether you
# want to install the CPU or GPU version of TensorFlow.
#
################################################################
# Basic packages used in many of the tutorials.

numpy
#scipy
#jupyter
matplotlib
pillow
#scikit-learn
opencv
flask
#colorama



################################################################
# TensorFlow v.2.1 and above include both CPU and GPU versions.

tensorflow

################################################################
# Some tutorials use other individual Python packages.
# Uncomment the relevant lines for the tutorials you want to run.

# gym[atari]  # Tutorial #16 on Reinforcement Learning.
# pandas      # Tutorial #23 on Time-Series Prediction.

################################################################
# PrettyTensor was used as the builder API for several of the
# earlier tutorials. PrettyTensor is apparently no longer being
# maintained and may not work with newer versions of TensorFlow.

# prettytensor

################################################################
